main: add handler to retrieve numbers for an image
I intend to use this handler for statistical things, ie. which
percentage of numbers are >1000, which percentage of numbers are < 10.
Depending on how many users the site has there could be a scaling issue
with returning all the numbers without a pagination system.... But I
think that can wait (famous last words #720)