main: random image getter
I am worried about performance of this handler (getRandomImageHandler),
as it retrieves all of the images, checks the length and then soley
returns an object within a random index. This could potentially be sped
up with some mongo-foo... But premature optimisation is the spawn of all
evil... I'm just kidding /dev/null is.